cleaned up example don HeapView, eliminating DataHeap (a repetition)
