+++ Examples/src/HeapView/mainHeap.cpp
@@ -3,6 +3,24 @@
 //
 #include "heapView.hpp"
 #include <iostream>
+#include <iomanip>
+// A small function to print a heap stored in a vector
+void printHeap(auto const & heap, std::size_t idx = 0, std::size_t indent = 0) {
+    if (idx < heap.size()) {
+        printHeap(heap, 2*idx + 2, indent + 4);
+        std::cout << std::setw(indent) << ' ';
+        std::cout << heap[idx] << "\n";
+        printHeap(heap, 2*idx + 1, indent + 4);
+    }
+}
+void printHeapView(auto const & heapView) {
+  std::vector<double> heap;
+  for(auto i = 0u; i < heapView.size(); ++i)
+    {
+      heap.push_back(heapView.heapValue(i));
+    }
+  printHeap(heap);
+}
 
 int
 main()
@@ -18,21 +36,15 @@ main()
     std::cout << "heapView top = " << heapView.top() << std::endl;
     std::cout << "heapView size = " << heapView.size() << std::endl;
     std::cout << "current values in the heap = " << std::endl;
-    for(auto i = 0u; i < heapView.size(); ++i)
-      {
-        std::cout << heapView.heapValue(i) << " ";
-      }
+    printHeapView(heapView);
     std::cout << std::endl;
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
-    auto [index, value] = heapView.pop();
+    auto [index, value] = heapView.popPair();
     std::cout << "heapView pop = " << index << " " << value << std::endl;
     std::cout << "heapView size = " << heapView.size() << std::endl;
     std::cout << "current values in the heap = " << std::endl;
-    for(auto i = 0u; i < heapView.size(); ++i)
-      {
-        std::cout << heapView.heapValue(i) << " ";
-      }
+    printHeapView(heapView);
     std::cout << std::endl;
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
@@ -41,10 +53,7 @@ main()
     std::cout << "heapView top = " << heapView.top() << std::endl;
     std::cout << "heapView size = " << heapView.size() << std::endl;
     std::cout << "current values in the heap = " << std::endl;
-    for(auto i = 0u; i < heapView.size(); ++i)
-      {
-        std::cout << heapView.heapValue(i) << " ";
-      }
+    printHeapView(heapView);
     std::cout << std::endl;
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
@@ -53,10 +62,7 @@ main()
     std::cout << "heapView top = " << heapView.top() << std::endl;
     std::cout << "heapView size = " << heapView.size() << std::endl;
     std::cout << "current values in the heap = " << std::endl;
-    for(auto i = 0u; i < heapView.size(); ++i)
-      {
-        std::cout << heapView.heapValue(i) << " ";
-      }
+    printHeapView(heapView);
     std::cout << std::endl;
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
@@ -65,10 +71,7 @@ main()
     std::cout << "heapView top = " << heapView.top() << std::endl;
     std::cout << "heapView size = " << heapView.size() << std::endl;
     std::cout << "current values in the heap = " << std::endl;
-    for(auto i = 0u; i < heapView.size(); ++i)
-      {
-        std::cout << heapView.heapValue(i) << " ";
-      }
+    printHeapView(heapView);
     std::cout << std::endl;
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
@@ -81,21 +84,15 @@ main()
     std::cout << "heapView top = " << heapView.top() << std::endl;
     std::cout << "heapView size = " << heapView.size() << std::endl;
     std::cout << "current values in the heap = " << std::endl;
-    for(auto i = 0u; i < heapView.size(); ++i)
-      {
-        std::cout << heapView.heapValue(i) << " ";
-      }
+    printHeapView(heapView);
     std::cout << std::endl;
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
-    auto [index, value] = heapView.pop();
+    auto [index, value] = heapView.popPair();
     std::cout << "heapView pop = " << index << " " << value << std::endl;
     std::cout << "heapView size = " << heapView.size() << std::endl;
     std::cout << "current values in the heap = " << std::endl;
-    for(auto i = 0u; i < heapView.size(); ++i)
-      {
-        std::cout << heapView.heapValue(i) << " ";
-      }
+    printHeapView(heapView);
     std::cout << std::endl;
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
@@ -104,10 +101,7 @@ main()
     std::cout << "heapView top = " << heapView.top() << std::endl;
     std::cout << "heapView size = " << heapView.size() << std::endl;
     std::cout << "current values in the heap = " << std::endl;
-    for(auto i = 0u; i < heapView.size(); ++i)
-      {
-        std::cout << heapView.heapValue(i) << " ";
-      }
+    printHeapView(heapView);
     std::cout << std::endl;
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
@@ -116,10 +110,7 @@ main()
     std::cout << "heapView top = " << heapView.top() << std::endl;
     std::cout << "heapView size = " << heapView.size() << std::endl;
     std::cout << "current values in the heap = " << std::endl;
-    for(auto i = 0u; i < heapView.size(); ++i)
-      {
-        std::cout << heapView.heapValue(i) << " ";
-      }
+    printHeapView(heapView);
     std::cout << std::endl;
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
@@ -128,35 +119,29 @@ main()
     std::cout << "heapView top = " << heapView.top() << std::endl;
     std::cout << "heapView size = " << heapView.size() << std::endl;
     std::cout << "current values in the heap = " << std::endl;
-    for(auto i = 0u; i < heapView.size(); ++i)
-      {
-        std::cout << heapView.heapValue(i) << " ";
-      }
+    printHeapView(heapView);
     std::cout << std::endl;
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
     std::cout<< "Emptying heap\n";
     auto currentSize=heapView.size();
     std::vector<std::size_t> currentHeapElements;
-    for (auto i=0;i<currentSize;++i)
+    for (auto i=0ul;i<currentSize;++i)
     {
-     auto where= heapView.pop();
+     auto where= heapView.popPair();
      currentHeapElements.emplace_back(where.first);
     }
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
-    std::cout<< "Putting back with update\n";
-    for (auto i=0;i<currentSize;++i)
+    std::cout<< "Putting elements back with update\n";
+    for (auto i=0ul;i<currentSize;++i)
     {
       auto k = currentHeapElements[i];
       heapView.update(k,heapView[k]);
     }
     std::cout << "heapView size = " << heapView.size() << std::endl;
     std::cout << "current values in the heap = " << std::endl;
-    for(auto i = 0u; i < heapView.size(); ++i)
-      {
-        std::cout << heapView.heapValue(i) << " ";
-      }
+    printHeapView(heapView);
     std::cout << std::endl;
     std::cout << "Heap is sane? " << std::boolalpha << heapView.check()
               << std::endl;
